Group adt adjesent types into `ty::adt`
